Add tiled image to viewer inside canvasObject
